# src/ownsms/openapi.yaml
...
openapi: 3.1.0
info:
  title: ownsms API
  version: "0.3.1"
  description: >
    Send SMS through your own phone's SIM via API. This is the implemented contract of the
    `ownsms` Django package (self-host). The cloud (ownsms.uz) and a future FastAPI deployment
    share the same contract.

    Two credentials: an **API key** (`osk_...`, developer → send/manage) and a **device token**
    (Android device → poll/report). Both are sent as `Authorization: Bearer <token>`.
servers:
  - url: https://sms.omadli.uz/api/v1
    description: Live demo (self-host reference deployment)
  - url: https://api.ownsms.uz/api/v1
    description: Cloud
  - url: "{base}/api/v1"
    description: Self-host (django-ownsms / ngrok / localhost)
    variables:
      base: { default: "http://localhost:8000" }
security:
  - apiKey: []
...
paths:
...
  /campaigns/{id}:
    get:
      tags: [Campaigns]
      summary: Campaign status + progress
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/Campaign" } } } }
  /campaigns/{id}/{action}:
    post:
      tags: [Campaigns]
      summary: pause | resume | cancel
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: action, in: path, required: true, schema: { type: string, enum: [pause, resume, cancel] } }
      responses:
        "200": { description: OK, content: { application/json: { schema: { $ref: "#/components/schemas/CampaignRef" } } } }
...
components:
  securitySchemes:
    apiKey: { type: http, scheme: bearer, description: "Developer API key (osk_...)" }
    deviceToken: { type: http, scheme: bearer, description: "Device token (issued at registration)" }
  schemas:
...
    CampaignRef:
      type: object
      properties:
        id: { type: string, example: camp_7f3a }
        status: { type: string, enum: [scheduled, running, paused, completed, canceled] }
        total: { type: integer }
    Campaign:
      allOf:
        - $ref: "#/components/schemas/CampaignRef"
        - type: object
          properties:
            progress:
              type: object
              properties:
                queued: { type: integer }
                sending: { type: integer }
                sent: { type: integer }
                delivered: { type: integer }
                failed: { type: integer }
                expired: { type: integer }
                canceled: { type: integer }
...